Adding some testing for time_deriv_RHS (but there a strange error with checking the looping forward indices; GTest tells me that -7160 does not equal -7160 and so a test fails, but the correct value is actually being returned...)

+++ source/KdV_main.cxx
@@ -1,10 +1,19 @@
+/**
+* The main function for the KdV solver code Emily and I are writing. This
+* code is repsonsible for stating initial parameters and calling the support
+* functions.
+*
+* Author: James "Andy" Edmond
+* Date: March 29, 2020
+*/
+
 #include "KdV.h"
 
 int main() {
   
   vector arr = xt::empty<double>({5});
   for (int i = 0; i < arr.size(); i++) {
-    arr(i) = float(i);
+    arr(i) = double(i);
   }
 
   double step = 0.1;
